+++ app/login/page.tsx
@@ -1,26 +1,15 @@
-'use client';
+import type { Metadata } from 'next';
+import LoginForm from '@/components/Auth/LoginForm';
 
-import { useState } from 'react';
-import { useAuth } from '@/context/Auth';
+export const metadata: Metadata = {
+  title: 'Login to Art Wise AI - Your Personal Art Guide',
+  description:
+    'Join Art Wise AI to explore artworks through AI-powered conversations. Get instant insights about art history, techniques, and cultural significance.',
+  keywords:
+    'art analysis app, AI art guide, artwork exploration, art education technology, art appreciation tool',
+};
 
 export default function LoginPage() {
-  const [email, setEmail] = useState('');
-  const [status, setStatus] = useState<'idle' | 'loading' | 'success' | 'error'>('idle');
-  const [errorMessage, setErrorMessage] = useState('');
-  const auth = useAuth();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setStatus('loading');
-    try {
-      await auth.sendSignInEmail(email);
-      setStatus('success');
-    } catch (error) {
-      setStatus('error');
-      setErrorMessage(error instanceof Error ? error.message : 'Failed to send login link');
-    }
-  };
-
   return (
     <div className="min-h-screen bg-gray-50">
       {/* Hero Section */}
@@ -28,127 +17,7 @@ export default function LoginPage() {
         <div className="max-w-7xl mx-auto">
           <div className="relative z-10 pb-8 bg-gray-50 sm:pb-16 md:pb-20 lg:w-full lg:pb-28 xl:pb-32">
             <main className="mt-10 mx-auto max-w-7xl px-4 sm:mt-12 sm:px-6 lg:mt-16 lg:px-8">
-              <div className="grid grid-cols-1 lg:grid-cols-2 gap-8 items-center">
-                {/* Left Column - Text Content */}
-                <div>
-                  <h1 className="text-4xl tracking-tight font-extrabold text-gray-900 sm:text-5xl md:text-6xl">
-                    <span className="block">Discover Art with</span>
-                    <span className="block text-blue-600">AI Companion</span>
-                  </h1>
-                  <p className="mt-3 text-base text-gray-500 sm:mt-5 sm:text-lg sm:max-w-xl sm:mx-auto md:mt-5 md:text-xl lg:mx-0">
-                    Explore artworks like never before with our AI-powered art companion. Get
-                    detailed insights, historical context, and engage in natural conversations about
-                    any artwork through text or voice chat.
-                  </p>
-
-                  {/* Feature List */}
-                  <div className="mt-8">
-                    <h2 className="text-lg font-semibold text-gray-900 mb-4">Key Features:</h2>
-                    <ul className="space-y-4">
-                      <li className="flex items-start">
-                        <svg
-                          className="h-6 w-6 text-green-500 mr-2"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                          stroke="currentColor"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={2}
-                            d="M5 13l4 4L19 7"
-                          />
-                        </svg>
-                        <span>Instant artwork analysis and insights</span>
-                      </li>
-                      <li className="flex items-start">
-                        <svg
-                          className="h-6 w-6 text-green-500 mr-2"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                          stroke="currentColor"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={2}
-                            d="M5 13l4 4L19 7"
-                          />
-                        </svg>
-                        <span>Voice and text chat with AI art historian</span>
-                      </li>
-                      <li className="flex items-start">
-                        <svg
-                          className="h-6 w-6 text-green-500 mr-2"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                          stroke="currentColor"
-                        >
-                          <path
-                            strokeLinecap="round"
-                            strokeLinejoin="round"
-                            strokeWidth={2}
-                            d="M5 13l4 4L19 7"
-                          />
-                        </svg>
-                        <span>Audio descriptions in multiple languages</span>
-                      </li>
-                    </ul>
-                  </div>
-                </div>
-
-                {/* Right Column - Login Form & Video */}
-                <div className="space-y-8">
-                  {/* Video Placeholder */}
-                  <div className="aspect-w-16 aspect-h-9 bg-gray-200 rounded-lg overflow-hidden">
-                    <div className="flex items-center justify-center h-full">
-                      <span className="text-gray-500">Demo video coming soon</span>
-                    </div>
-                  </div>
-
-                  {/* Login Form */}
-                  <div className="bg-white p-8 rounded-lg shadow-md">
-                    <h2 className="text-2xl font-bold text-gray-900 mb-6">Get Started</h2>
-                    <form onSubmit={handleSubmit} className="space-y-4">
-                      <div>
-                        <label htmlFor="email" className="block text-sm font-medium text-gray-700">
-                          Email address
-                        </label>
-                        <input
-                          type="email"
-                          id="email"
-                          value={email}
-                          onChange={(e) => setEmail(e.target.value)}
-                          className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-blue-500 focus:ring-blue-500"
-                          placeholder="Enter your email"
-                          required
-                        />
-                      </div>
-                      <button
-                        type="submit"
-                        disabled={status === 'loading'}
-                        className="w-full flex justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-blue-600 hover:bg-blue-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-blue-500 disabled:opacity-50"
-                      >
-                        {status === 'loading' ? 'Sending...' : 'Sign in with Email'}
-                      </button>
-                    </form>
-
-                    {/* Status Messages */}
-                    {status === 'success' && (
-                      <div className="mt-4 p-4 bg-green-50 rounded-md">
-                        <p className="text-sm text-green-700">
-                          Check your email for the login link!
-                        </p>
-                      </div>
-                    )}
-                    {status === 'error' && (
-                      <div className="mt-4 p-4 bg-red-50 rounded-md">
-                        <p className="text-sm text-red-700">{errorMessage}</p>
-                      </div>
-                    )}
-                  </div>
-                </div>
-              </div>
+              <LoginForm />
             </main>
           </div>
         </div>
